Перенес метод удаления рецепта в модуль RecipeDetails
