+++ CookingWebsite/CookingWebsite/ClientApp/src/app/pages/recipe-details/recipe-details.component.ts
@@ -94,7 +94,7 @@ export class RecipeDetailsComponent {
   public async deleteRecipe(){
     let respose = await this.recipeDetailsService.deleteRecipe(this.route.snapshot.params.id);
 
-    if (respose)
+    if (respose != null)
       this.router.navigate(['/recipes']);
   }
 }
