--- CookingWebsite/CookingWebsite/Modules/RecipeDeleteModule/RecipeDeleteBindings.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-namespace CookingWebsite.Modules.RecipeDeleteModule
-{
-    public static class RecipeDeleteBindings
-    {
-        public static IServiceCollection AddRecipeDeleteModule(this IServiceCollection services)
-        {
-            return services;
-        }
-    }
-}
