+++ CookingWebsite/CookingWebsite/Modules/RecipeDetailsModule/RecipeDetailsController.cs
@@ -1,4 +1,5 @@
-﻿using CookingWebsite.Domain.Repositories;
+﻿using CookingWebsite.Domain;
+using CookingWebsite.Domain.Repositories;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Security.Claims;
@@ -7,19 +8,25 @@ using System.Threading.Tasks;
 namespace CookingWebsite.Modules.RecipeDetailsModule
 {
     [ApiController]
-    [Route("api/recipes/{recipeId}/details")]
+    [Route("api/recipes/{recipeId}")]
     public class RecipeDetailsController : ControllerBase
     {
+        private readonly IUnitOfWork _unitOfWork;
         private readonly IRecipeRepository _recipeRepository;
         private readonly IRecipeLikeRepository _recipeLikeRepository;
 
-        public RecipeDetailsController(IRecipeRepository recipeRepository, IRecipeLikeRepository recipeLikeRepository)
+        public RecipeDetailsController(
+            IUnitOfWork unitOfWork,
+            IRecipeRepository recipeRepository,
+            IRecipeLikeRepository recipeLikeRepository
+            )
         {
+            _unitOfWork = unitOfWork;
             _recipeRepository = recipeRepository;
             _recipeLikeRepository = recipeLikeRepository;
         }
 
-        [HttpGet]
+        [HttpGet("details")]
         public async Task<RecipeDetailsDto> GetRecipeDetails([FromRoute] int recipeId)
         {
             var recipeDetails = await _recipeRepository.GetById(recipeId);
@@ -30,6 +37,16 @@ namespace CookingWebsite.Modules.RecipeDetailsModule
                 userLikes,
                 Convert.ToInt32(User.FindFirstValue(Claims.UserId))
                 );
+        }
+
+        [HttpPost("delete")]
+        public async Task DeleteRecipe([FromRoute] int recipeId)
+        {
+            var recipe = await _recipeRepository.GetById(recipeId);
+
+            _recipeRepository.Delete(recipe);
+
+            await _unitOfWork.Commit();
         }
     }
